requirements.txt
pytest
pillow
PyOpenGL
pyopengltk
glfw